Correcting a bug when label_get is after the label_set

+++ asm/src/write/args_manager.c
@@ -39,8 +39,8 @@ int get_arg_value(char *arg, int *pos, assembly_data_t *datas)
 			my_puterror("Cannot find label !\n");
 			return (0);
 		}
-		printf("%u\n", (unsigned int)(label_pos - *pos));
-		return ((unsigned int)(label_pos - *pos));
+		printf("%i\n", (short)((unsigned short)(label_pos - *pos)));
+		return ((unsigned short)(label_pos - *pos));
 	}
 	if (arg[0] < '0' || arg[0] > '9')
 		index = 1;
